+++ crates/rooch-sequencer/src/proxy/mod.rs
@@ -48,7 +48,7 @@ impl SequencerProxy {
 
     pub async fn get_transaction_sequence_info_mapping_by_order(
         &self,
-        tx_orders: Vec<u128>,
+        tx_orders: Vec<u64>,
     ) -> Result<Vec<Option<TransactionSequenceInfoMapping>>> {
         self.actor
             .send(GetTxSequenceInfoMappingByOrderMessage { tx_orders })
@@ -66,7 +66,7 @@ impl SequencerProxy {
 
     pub async fn get_transaction_sequence_infos(
         &self,
-        orders: Vec<u128>,
+        orders: Vec<u64>,
     ) -> Result<Vec<Option<TransactionSequenceInfo>>> {
         self.actor
             .send(GetTxSequenceInfosMessage { orders })
